requirements.txt
cffi==2.0.0
dearpygui==2.0.0
numpy==2.3.1
pycparser==2.23
pyserial==3.5
sounddevice==0.5.2
soundfile==0.13.1
